Rework contribution guideline
Content of `CONTRIBUTING.md` is moved to `contribution.md`.
This is required because CONTRIBUTING has special handling
and can be referenced only by full URL with the site name.
And even in this case, it is not rendered as markdown,
but as plain text.

CONTRIBUTING.md is kept for compatibility reasons.
Many of our projects are referring to it.
So corresponding text is placed inside of it.

Add reference to new `contribution.md` from README.md

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Reviewed-by: Andrii Chepurnyi <andrii_chepurnyi@epam.com>